requirements.txt
fastapi
jinja2 
uvicorn
pydantic
pydantic[email]
pymongo
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-sdk[console]
opentelemetry-exporter-otlp
opentelemetry-instrumentation-pymongo
opentelemetry-exporter-otlp-proto-grpc
python-multipart 
bcrypt